backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
psycopg2-binary>=2.9.9
alembic>=1.13.0
python-decouple>=3.8
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
